+++ maidr/core/plot/grouped_barplot.py
@@ -24,7 +24,7 @@ class GroupedBarPlot(
         grouped_ax_schema = {
             MaidrKey.X.value: self.ax.get_xlabel(),
             MaidrKey.Y.value: self.ax.get_ylabel(),
-            MaidrKey.FILL.value: self.extract_level(self.ax, MaidrKey.FILL),
+            MaidrKey.FILL.value: self.ax.get_title(),
         }
         return self.merge_dict(base_ax_schema, grouped_ax_schema)